Overwrite the CI_Email::send() function
This way we can prevent CI from sending emails in the testing environment. This was a problem because Github Actions offered a way to send emails using the default mail command that were actually being send.

In development, all mail is send to mailhog so the output from the email can be tested there.